编译警告

+++ media/src/h263/h263codec.h
@@ -46,7 +46,7 @@ public:
 	virtual int Decode(BYTE *in,DWORD len);
 	virtual int GetWidth()	{return ctx->width;};
 	virtual int GetHeight()	{return ctx->height;};
-	virtual int GetFPS() {return round(1 / av_q2d(ctx->time_base));};
+	virtual int GetFPS() {return (int)round(1 / av_q2d(ctx->time_base));};
 	virtual BYTE* GetFrame(){return (BYTE *)frame;};
 	virtual DWORD GetFrameSize(){return frameSize;};
 	virtual int GetBitRate(){return ctx->bit_rate;};
@@ -117,7 +117,7 @@ public:
 	virtual int Decode(BYTE *in,DWORD len);
 	virtual int GetWidth()	{return ctx->width;};
 	virtual int GetHeight()	{return ctx->height;};
-	virtual int GetFPS() {return round(1 / av_q2d(ctx->time_base));};
+	virtual int GetFPS() {return (int)round(1 / av_q2d(ctx->time_base));};
 	virtual BYTE* GetFrame(){return (BYTE *)frame;};
 	virtual DWORD GetFrameSize(){return frameSize;};
 	virtual int GetBitRate(){return ctx->bit_rate;};
